#16 move GMPISubClientIndividual from geneva-individuals to geneva, because it is not a complete individual but an abstract base class for user defined individuals
