+++ include/geneva-individuals/CMakeLists.txt
@@ -9,13 +9,6 @@ SET ( SAMPLEINDIVIDUALINCLUDES
 	GTestIndividual3.hpp
 )
 
-# Add MPI specific files
-IF(GENEVA_BUILD_WITH_MPI_CONSUMER)
-	SET(SAMPLEINDIVIDUALINCLUDES
-			${SAMPLEINDIVIDUALINCLUDES}
-			GMPISubClientIndividual.hpp)
-ENDIF()
-
 SET ( SAMPLEINDIVIDUALCONFIGS
 	# No config files at the moment
 )
