+++ src/geneva-individuals/CMakeLists.txt
@@ -45,14 +45,6 @@ SET ( SAMPLEINDIVIDUALSOURCES
 	GTestIndividual3
 )
 
-# Add MPI specific files
-IF(GENEVA_BUILD_WITH_MPI_CONSUMER)
-	SET(SAMPLEINDIVIDUALSOURCES
-			${SAMPLEINDIVIDUALSOURCES}
-			GMPISubClientIndividual)
-ENDIF()
-
-
 set_source_files_properties(
     ${SAMPLEINDIVIDUALSOURCES}
     PROPERTIES
