+++ src/geneva/CMakeLists.txt
@@ -99,6 +99,13 @@ SET ( OPTIMIZATIONSOURCES
 IF(GENEVA_BUILD_WITH_MPI_CONSUMER)
 	SET(OPTIMIZATIONSOURCES
 			${OPTIMIZATIONSOURCES}
+			GMPISubClientIndividual)
+ENDIF()
+
+# Add MPI specific files
+IF(GENEVA_BUILD_WITH_MPI_CONSUMER)
+	SET(OPTIMIZATIONSOURCES
+			${OPTIMIZATIONSOURCES}
 			GMPISubClientOptimizer)
 ENDIF()
